### 2020-05-22 #2209 @whedon
[PRE REVIEW]: DriveML: An R Package for Driverless Machine Learning
**Submitting author:** @sayanddude (<a href="http://orcid.org/0000-0002-6477-0376">Sayan Putatunda</a>)
**Repository:** <a href="https://github.com/daya6489/DriveML" target ="_blank">https://github.com/daya6489/DriveML</a>
**Version:** v0.1
