### 2020-05-23 #2209 @sayanddude
Hi, updated the paper.bib file in the github repo and re-generating the pdf to resolve the issue of "Invalid DOIs"
