### 2020-06-03 #2278 @mirca
Dear @sayanddude, thank you for submitting your paper to JOSS.  I've open several issues on the repository https://github.com/daya6489/DriveML regarding my review on the manuscript. Please, take a look at your convenience.
