### 2020-06-05 #2278 @sayanddude
> @sayanddude You can use http://whedon.theoj.org/ to test your paper compilation.

Thanks a lot!
