### 2020-06-05 #2278 @sayanddude
Hi @mirca , Thanks for your comments! I have addressed all the comments, updated the manuscript and have generated the latest pdf above. Please find below my response to each of the comments. Also, I have separately responded to the comments posted in Github issues at- https://github.com/daya6489/DriveML/issues
--------------------------------------------
**1)	License file is inconsistent with the license reported in the paper**
Response- Thank you for pointing this out! We have now updated the Github repo with a GNU General Public License (GPL-3) and have mentioned the same in the paper as well.

**2)	Installation-** The installation method reported is conditioned on the acceptance of the package into the Comprehensive R Archive Network (CRAN). Therefore, the package…………..
Response- Thanks for your comment! The DriveML package is now accepted in CRAN and it is available at https://cran.r-project.org/web/packages/DriveML/index.html. So, we have updated the “Installation” section in the Github repo with both ways to install the package i.e. via CRAN and devtools.

**3)	Affiliation**
Response- Thanks for pointing this out! EDA AA & DS CoE stands for Enterprise Data Analytics- Advanced Analytics and Data Sciences Center of Excellence. We have updated the paper with the full form.

**4)	Comments on the Summary section**
Response- Thanks for your comments! We have provided an appropriate reference i.e. (Tuggener et al., 2019) for AutoML as suggested. We have also worked on the other two comments and have rewritten the specific lines.

**5)	Comments on Key Functionality section**- "Figure 2" is more like a table, I would suggest to refer to it as "Table 1".
Response- Thanks for your comment! We have replaced Figure 2 (of the earlier version) with Table 1 in the latest version of the paper.

**6)	Comments on Illustration section**
Response- Thanks for your comments! We agree about your specific comments for the illustration section and so we have rewritten this section entirely. The focus is now not on comparison of methods but how to practically use the DriveML package. Here we focus on the functionalities of DriveML viz., exploratory overview of the data, automated data preparation, automated machine learning model building and generating report in HTML. We use a publicly available dataset and write the corresponding R codes as well.

Regarding your comments on the unique features of DriveML mentioned in Table 1- "identifying pattern of missing values" and "HTML report using rmarkdown", we have updated the “Key Functionality” section of the paper where we have explained how the “autoMAR” function works and have provided a sample HTML report/vignette generated by DriveML that is published in CRAN.

**7)	Comments on Quality of Writing**- There are several grammatical issues with the manuscript, please take your time reading the manuscript thoroughly.
Response- Thanks for your comment! We have performed a thorough proof reading of the paper and also used tools like Grammarly. We have corrected all the grammatical and typographical errors in the updated version of the manuscript.
-------------------------------------------------------------------
Please let me know in case of any issues. Thanks!
