### 2020-06-18 #2278 @pat-s
# Review of DriveML

The package is practically a bundle several wrapper scripts around existing functionality conducting opinionated preprocessing and model selection steps.
Further it has a questionable value for other users because the motivations for many choices are missing.
The main motivation proofing that this package performs better than other packages is flawed because the scope of the packages listed is different than the one of {DriveML}.

Also, leaving choices for intermediate steps to the user is usually done by design by the underlying packages (like mlr or caret) and not a feature that is missing, hence the usefulness of such an automation can be questioned per se.
I haven't opened any issues in the repo so far because many of my points listed below are on the general side and will interact with each other, possibly invalidating others.

I'd rate the R package implementation on a technical side as "basic".
The test suite is very minimal.
The package comes with a long list of (recursive) dependencies which can be seen as critical.
R packages should aim to be minimal, especially when they should be used by others.
In addition, the packages makes use of two deprecated ML frameworks (caret & mlr).
For both, successors (tidymodels & mlr3) exist since quite some time.

As pointed out by @mirca, there are also several grammar and wording issues that would need to be addressed to avoid changing the meaning of certain sentences/paragraphs to something undesired.

In summary, I do not think that this package adds any novelty or sophisticated approaches to automate (parts) of a machine learning workflow effectively. In contrast, I fear that such low-quality, opinionated automation packages published in a journal will result in questionable ML studies/inferences by others with the claim of using a package that is a "reviewed" journal paper.
Hence, my suggestion would be to reject it.
However, as JOSS has the policy of not rejecting submissions per se, I would then label it as "major revisions".
I do not think that one round of revisions would be enough to address all points in a sufficient manner.
Hence, I would like to opt-out / resign from a possible multi-round review of this submission.
Nevertheless, thanks for submitting this work and I hope that my points listed below are still considered as helpful by the authors.

Detailed comments below

<summary>Details</summary>

## Package Structure

- Roxygen version is outdated
- Does the "heart" dataset necessarily need to come with this package or is it already available in other packages?
- Why was the "hear" dataset chosen instead of one of the common ML datasets in R available in various packages?
- Almost no git history, it seem the package was uploaded to git right before submission

### Authors

The DESCRIPTION file list way more authors than there are contributions listed in the GitHub repo.

## Installation

The install command of the dev version could be improved:

- devtools should be replaced by remotes as the former just wraps the latter but comes with way more dependencies
- No need to set `ref = master` as this is already the default
- The user should not be forced to install the vignettes via `build_vignettes = TRUE`

### README

- It is good practice to put a vignette into a pkgdown page or keep it in a .Rmd file rather than putting (parts?) of it into the README. This creates unneeded redundancy.
- It is not good practice to put `install.packages()` calls into a vignette/README (e.g. `install.packages("DriveML")` as these for a reinstallation of packages every time, even this is often not necessary.
- desease → disease j

### Dependencies

- Heavy dependencies with recursive heavy dependencies:
    - DriveML → no source coude available?
    - SmartEDA

# Functions

- `ExpData(data = heart, type = 1)`
    - Categorizes variables into groups. Some of these are R-specific and not general variable groups. For example, only in R there are "factor" vectors. However, these are just nominal/distinct variables in a ML sense. Hence, "text", "factor" and "logical" variables are all the same. Applying this categorization might cause users to assume that there are fundamental implications for ML models (which there are not). I suggest removing those special categories.
    - What are "unique variables"?
    - `%. of Variables having >90% missing cases 0% (0)` → why 90%?
    - `11 %. of Variables having <50% missing cases 0% (0)`
    `12 %. of Variables having >50% missing cases 0% (0)`

         How can both be zero here? They should sum up to 100%.

    - The column name "Obs" is misleading as observations are rows in a dataset. Here they are used for counts of any kind. Suggesting to rename the column name.

- `ExpData(data = heart, type = 2)`
    - categorizes between integers and numerics now, whereas `type = 1` does not
    - The first column name is undescriptive ("S.no")

- The help file of `ExpData` needs a spell check
- `ExpNumStat()`
    - Returned DF has different col names than `ExpData` for the same content ("Per_of_Missing" vs. "`% of Missing`"
    - The argument descriptions randomly start Upper case or lower case, the same goes for the variable names
    - The author name listed in the help page should be a fully qualified person name or removed (currently "dubrangala")

- `ExpNumViz()` : the code in the README does not run. It contains arguments that are not (anymore) present in the function
    - opened 17(!) external graphic windows which need to be closed individually. This is very tedious and I also do not see any value in this function. These cannot be compared to each other in different graphic windows, the scales are different, etc.
    - The function does not mention how the plots are combined and composed internally. The user can guess that ggplot2 is used but is somewhat suprised by the result

    `autoDataprep()`

    - returns hundreds of lines. At least the printer (list of 12) needs to be adjusted. There is a dedicated printer function `printautoDataprep()` which should be the S3 printing method for class `autoDataprep` (which has an inconsistent camel case naming scheme)
    - The arguments have types here whereas other functions have not
    - Presumably the user should continue with `foo$master_data` - this is not clear from the documentation

    `autoMLmodel()`

    - The default settings in the README example take very long time (100 iters, 20 varImp iterations). Examples should
    - The printer needs to be adjusted#

    `autoMLReport()` errors

    # Wording

    - "most difficult machine learning functions" → there is no ML function, these are algorithms. None is "difficult", "complex" would be a more appropriate wording. If "function" is meant as "tasks" or "steps" than this should also be renamed.
    - "DriveML R package has four unique functionalities" → unique is wrong word here
    - The grammer and wording feels very sloppy, no grammer or spell check was done. The same applies to the help pages.

    ## Tests

    - Test suite is minimal, no test coverage is provided

    ## Manuscript

    - Review unnecessary uses of "the"
    - Feature engineering is not a necessary step in ML, purely optional
    - Data Preparation and Feature Engineering can be summarised as "Data preprocessing"
    - "DriveML performs the best across different parameters." → which parameters?
    - How does DriveML address reproducibility in any way?
    - Is there a convention how to format R packages and functions?
    - Reading the "key functionality" one could have the impression that only (b) is done via mlr and everything else is functionality from DriveML itself?
    - Why were two implementations of randomForest integrated?
    - Fig.1: caret and mlr do not do any automated ML, they provide building blocks for creating ML pipelines and do not enforce and opionionated defaults (as DriveML does).
    - Fig.1: The comparison made is not fair as many of the mentioned packages do not make any choice about meaningful settings for tuning and or preprocessing
    - "Thus, Table 1 shows that DriveML is better than almost all the other packages available in CRAN/Github" → This is a very strong statement which I would vote against.
    - The paper contains more print output than text and is also very sparse on references
    - The coding style is not consitent with respect to whitspaces around operators
    - No summary/discussion of the results

    ### Grammar

    Suggestions for parts that could be rephrased:

    - "we introduce a new package i.e. DriveML" (Summary)
    - reduce the use of the word "pillar", maybe replace it completely - I don't think it is the correct word in this context
    - Github → GitHub
    - "reduce developer’s errors"
    - Watch spaces after hyphens
    - Watch spaces after brackets
    - "HTML slide/vignette" → define what it is. It can't be both.
    - "This report follows the industry and academia’s best practices." → What are the best practices and how are they defined?
    - "sample HTML report/vignette"
