### 2020-06-22 #2278 @sayanddude
@pat-s Thank you for your detailed comments! We have started working on addressing the issues you raised on Package Structure, Installation, README , Dependencies, Functions, Tests, Wording, Manuscript and Grammar. We plan to complete this in a week or so and we will get back to you with the updated package and manuscript.
