+++ README.md
@@ -137,3 +137,4 @@ autoMLReport(mlobject = small_ml_random, mldata = small_data, op_file = "driveML
 
 ## Articles
 
+The pre-print version of the paper on DriveML is available at ArXiv at- https://arxiv.org/pdf/2005.00478.pdf. Also, currently the paper is under review at the Journal of Open Source Software (JOSS).
