### 2020-09-08 #2278 @sayanddude
Hi @pat-s  , Thanks for your comments! I have addressed all the comments and updated both the manuscript and the package. The latest pdf is generated above.  Please find below my response to each of your comments.
-------------------------------------------------------------------------------------------------------------------

**Package Structure**
•	Roxygen version is outdated- Thanks for pointing it out! We have updated it in the latest version of the package.
•	Does the "heart" dataset necessarily need to come with this package or is it already available in other packages?- The heart disease dataset is publicly available in the UCI machine learning repository. It is not available in any other package but it is bundled with our package.
•	Why was the "heart" dataset chosen instead of one of the common ML datasets in R available in various packages?- It is a standard and a commonly used publicly available dataset for binary classification.
•	Almost no git history, it seem the package was uploaded to git right before submission- Actually, we first uploaded everything to CRAN. Only for final publication, we commit it top github. That’s why no history till now Github. However, for this version one can find the history.

**Authors**
The DESCRIPTION file list way more authors than there are contributions listed in the GitHub repo.- All the listed authors have equally contributed in the development of the package and in the paper writing. It is just that we are sharing different responsibilities.

**Installation**
The install command of the dev version could be improved:
•	devtools should be replaced by remotes as the former just wraps the latter but comes with way more dependencies- Thanks for your comment! It is fixed.
•	No need to set ref = master as this is already the default- It is fixed.
•	The user should not be forced to install the vignettes via build_vignettes = TRUE- Thanks! It is fixed.

**README**
•	It is good practice to put a vignette into a pkgdown page or keep it in a .Rmd file rather than putting (parts?) of it into the README. This creates unneeded redundancy.- Thanks for pointing it out! We have addressed this.
•	It is not good practice to put install.packages() calls into a vignette/README (e.g. install.packages("DriveML") as these for a reinstallation of packages every time, even this is often not necessary.- Thanks! It’s done.
•	desease → disease j- We have corrected it.

**Dependencies**
•	Heavy dependencies with recursive heavy dependencies:
o	DriveML → no source coude available?- The source code is available in github.
o	SmartEDA- It’s fixed! There ius no longer a dependency on SmartEDA.

**Functions**
•	ExpData(data = heart, type = 1)- This is a function of the SmartEDA package and since there is no longer any dependency on this package, so this is no longer a concern as it is not relevant to the DriveML package!
•	ExpData(data = heart, type = 2)- This is a function of the SmartEDA package and since there is no longer any dependency on this package, so this is no longer a concern as it is not relevant to the DriveML package!
•	The help file of ExpData needs a spell check- This is a function of the SmartEDA package and since there is no longer any dependency on this package, so this is no longer a concern as it is not relevant to the DriveML package!
•	ExpNumStat()- This is a function of the SmartEDA package and since there is no longer any dependency on this package, so this is no longer a concern as it is not relevant to the DriveML package!
•	ExpNumViz() : This is a function of the SmartEDA package and since there is no longer any dependency on this package, so this is no longer a concern as it is not relevant to the DriveML package!
•	autoDataprep()- This is a relevant function for DriveML.
o	returns hundreds of lines. At least the printer (list of 12) needs to be adjusted. There is a dedicated printer function printautoDataprep() which should be the S3 printing method for class autoDataprep (which has an inconsistent camel case naming scheme)- Thanks for your comment! We have now reduced it to a list of 6.
o	The arguments have types here whereas other functions have not- It’s fixed!
o	Presumably the user should continue with foo$master_data - this is not clear from the documentation-- It’s fixed!
autoMLmodel()
o	The default settings in the README example take very long time (100 iters, 20 varImp iterations). Examples should- make it 10 iters- Thanks for pointing it out! We have now made it 10 iterations.
o	The printer needs to be adjusted#- It’s fixed
autoMLReport() errors- All errors are fixed.

**Wording**
o	"most difficult machine learning functions" → there is no ML function, these are algorithms. None is "difficult", "complex" would be a more appropriate wording. If "function" is meant as "tasks" or "steps" than this should also be renamed.- fixed
o	"DriveML R package has four unique functionalities" → unique is wrong word here
o	The grammer and wording feels very sloppy, no grammer or spell check was done. The same applies to the help pages.- We have done an extensive grammar and spell check for both the paper and the help files.

**Tests**
o	Test suite is minimal, no test coverage is provided- Thanks for your comment! We have added few more test tests.

**Manuscript**
o	Review unnecessary uses of "the"- Thanks! We have corrected it.
o	 Feature engineering is not a necessary step in ML, purely optional—It is one of the important components of AutoML as per He, Zhao, & Chu (2020)
o	Data Preparation and Feature Engineering can be summarised as "Data preprocessing"- Agreed!
o	"DriveML performs the best across different parameters." → which parameters?—Thanks for pointing it out! We meant “across different features”. It is corrected in the latest version of the manuscript.
o	How does DriveML address reproducibility in any way?- We have fixed a set seed command before running all the models. That should take care of the reproducibility part.
o	Is there a convention how to format R packages and functions?- We used “lintr” check to format functions, question, etc. We have seen that quite a few joss paper/packages used lintr checks as a standard.
o	Reading the "key functionality" one could have the impression that only (b) is done via mlr and everything else is functionality from DriveML itself?
o	Why were two implementations of randomForest integrated?- We just wanted to give the user multiple options. By the way, a random forest model developed using ranger is much faster than the one developed using the RandomForest package.
o	Fig.1: caret and mlr do not do any automated ML, they provide building blocks for creating ML pipelines and do not enforce and opionionated defaults (as DriveML does).- We have removed caret and mlr from Table 1 (it was Fig. 1 in the earlier version of the manuscript) in the latest version of the manuscript.
o	Fig.1: The comparison made is not fair as many of the mentioned packages do not make any choice about meaningful settings for tuning and or preprocessing—That is true! However, since tuning and preprocessing are an integral part of autoML so we have included them as desired characteristics for comparing different packages.
o	"Thus, Table 1 shows that DriveML is better than almost all the other packages available in CRAN/Github" → This is a very strong statement which I would vote against.- Thanks for pointing it out! We have modified the last paragraph of the Section – “Comparison of DriveML with other relevant R Packages”.
o	The paper contains more print output than text and is also very sparse on references- Thanks for your comment! We wanted to showcase the functions of DriveML and how to use them and that’s why the Illustration section has detailed R codes. In case of references, although there have been quite a few works reported in the literature for autoML however, we wanted to keep the paper short (as it is one of the guidelines of JOSS) and so have kept the most relevant references.
o	The coding style is not consitent with respect to whitspaces around operators- We have performed a  “lintr” check that will remove all the unnecessary whitespaces and thus makes the code standardized. We have used this in one of our earlier packages that was published by JOSS.
o	No summary/discussion of the results- The results are discussed in the Illustration section where we explain Figures 2, 3, 4 and 5.

**Grammar**
Suggestions for parts that could be rephrased:
o	"we introduce a new package i.e. DriveML" (Summary)—Replaced it with…  “we introduce a new package named "DriveML" for automated machine learning.”
o	reduce the use of the word "pillar", maybe replace it completely - I don't think it is the correct word in this context. ----- Thanks for pointing this out! We have replaced it with “components”.
o	Github → GitHub—Corrected!
o	"reduce developer’s errors" —Corrected!
o	Watch spaces after hyphens —Corrected!
o	Watch spaces after brackets —Corrected!
o	"HTML slide/vignette" → define what it is. It can't be both.- Thanks for pointing it out! We meant “Vignette” and have made the correction.
o	"This report follows the industry and academia’s best practices." → What are the best practices and how are they defined?- Agreed… it sounds too vague. So, we have removed this line from the latest version of the manuscript.
o	"sample HTML report/vignette"- Corrected!

-----------------------------------------------------------------------------------------------------
