### 2020-09-20 #2278 @mirca
@sayanddude thanks for addressing some of the issues pointed out on the review process.  I have opened issues at https://github.com/daya6489/DriveML with some of my suggestions.  The quality of the writing is still a major concern, in my view.
