+++ inst/rmd_template/report_tmp_v1.Rmd
@@ -36,14 +36,14 @@ mldata <- params$mldata
 
 **Overview of the data**
 
-```{r mdp1,warning=FALSE,eval=T,render=t1,echo=F}
+```{r mdp1,warning=FALSE,eval=T,echo=F}
 typ1 <- mldata$datasummary$type1
 kable(typ1)
 ```
 
 **Variable summary**
 
-```{r mdp2,warning=FALSE,eval=T,render=t1,echo=F}
+```{r mdp2,warning=FALSE,eval=T,echo=F}
 typ2 <- mldata$datasummary$type2
 paged_table(typ2)
 ```
@@ -59,13 +59,21 @@ print(mldata$var_list$MAR_col)
 ### Missing imputation
 
 ```{r mdp4, warning = FALSE, eval = T, echo = F}
+
+if(names(mldata$call) %in% "missimpute") {
+
 if(mldata$call$missimpute=="default") {
   mdd <- data.frame(Variable_type = c("Factor", "Integer", "Numeric", "Character"),
   Impute_method = c("Mode value", "Mean value", "Median value", "Constant value"))
   kable(mdd)
     } else {
     cat("Missing replaced with", "{", print(mldata$call$missimpute), "\n") }
-
+  } else 
+  {
+    mdd <- data.frame(Variable_type = c("Factor", "Integer", "Numeric", "Character"),
+  Impute_method = c("Mode value", "Mean value", "Median value", "Constant value"))
+  kable(mdd)
+  }
 ```
 
 ### Feature engineering
@@ -81,7 +89,7 @@ kable(ccd)
 ##### Outlier variable
 
 ```{r mdp6, warning = FALSE, eval = T, echo = F}
-print(mldata$outlier_summary)
+print(mldata$var_list$outlier_summary)
 ```
 
 ##### Interaction variables using multiplaction method
@@ -91,7 +99,7 @@ Calculated ineraction between two continuous variables using multiplication meth
 List of variables calculated are below
 
 ```{r mdp7, warning = FALSE, eval = T, echo = F}
-cc  <- mldata$overall_variable
+cc  <- mldata$var_list$overall_variable
 print(cc[grep("Inter_", cc)])
 ```
 
@@ -103,7 +111,7 @@ The frequent transformer counts each categorical variable in the data set
 List of variables calculated are below
 
 ```{r mdp8, warning = FALSE, eval = T, echo = F}
-cc  <- mldata$overall_variable
+cc  <- mldata$var_list$overall_variable
 print(cc[grep("Freq_t", cc)])
 ```
 
@@ -114,7 +122,7 @@ Date transformer retrieves any date values like month, year, quarter and week
 List of variables calculated are below
 
 ```{r mdp9, warning = FALSE, eval = T, echo = F}
-cc  <- mldata$overall_variable
+cc  <- mldata$var_list$overall_variable
 print(cc[grep("datet_", cc)])
 ```
 
@@ -123,7 +131,7 @@ print(cc[grep("datet_", cc)])
 One hot coding for all categorical variables
 
 ```{r mdp10, warning = FALSE, eval = T, echo = F}
-cc  <- mldata$var_list$Dummy_col
+cc  <- mldata$var_list$overall_variable
 print(cc)
 
 ```
@@ -133,7 +141,7 @@ print(cc)
 ##### Zero variance
 
 ```{r mdp11, warning = FALSE, eval = T, echo = F}
[…]
