+++ GOVERNANCE.md
@@ -1,0 +1,47 @@
+## Users
+ 
+People who use DriveML for their data analytics work are the most important members of the community; without these users, this project would have no purpose.
+ 
+Users are encouraged to participate in the life of the project and the community as much as possible. Common user activities include (but are not limited to):
+ 
+- evangelizing about the project
+- asking and answering on community forums
+- providing moral support 
+ 
+Users who continue to engage with the project and its community will often find themselves becoming more and more involved. 
+ 
+## Contributors
+ 
+Contributors interact with the project on GitHub by filing new issues, improving existing issues, or submitting pull requests. Anyone can become a contributor: there is no expectation of commitment to the project, no required set of skills, and no selection process. The only obligation is to follow the [code of conduct](CODE_OF_CONDUCT.md).
+ 
+Specific advice for contributing to the project can be found in
+[CONTRIBUTING.md](https://github.com/daya6489/DriveML/blob/master/CONTRIBUTING.md).
+ 
+## Core developers
+ 
+Core developers are collectively responsible for day-to-day development of the package, including responding to issues and reviewing pull requests. They are GitHub administrators and [package authors](https://github.com/daya6489/DriveML/blob/master/DESCRIPTION), which means that they have the ability to make changes to project code, and receive credit when others cite the package.
+ 
+While core developers can modify code directly, this ability is rarely used. Instead, changes are proposed as pull requests, and are only merged after they have been reviewed by at least one other core developer. Changes to the API (especially breaking changes) must also be approved by the project lead. 
+ 
+ 
+The core developers of DriveML are:
+ 
+* [Sayan Putatunda]( https://github.com/sayanddude )
+* [Kiran R](https://github.com/rkirana)
+* [Ravi Kondapalli](https://github.com/rpkon)
+ 
+All core developers are bound by the [code of conduct](CODE_OF_CONDUCT.md).
+ 
+## Project lead
+ 
+The project lead, [Dayananda Ubrangala](https://github.com/daya6489), is responsible for:
+ 
+* Setting, and clearly communication the strategic objectives of the project.
+* Mediating any conflicts amongst the core developers.
+* Ensuring that the project survives in the long term.
+ 
+The project lead is bound by the [code of conduct](CODE_OF_CONDUCT.md).
+ 
+# Decision-making process
+ 
+This project makes decisions according to a consensus model where suggestions are considered and discussed between the community and core developers, typically in GitHub issues. Where consensus cannot be reached, the project lead's word is final. If the community questions a decision, the project lead may review it and either uphold or reverse it.
