### 2020-09-22 #2278 @sayanddude
> @sayanddude thanks for addressing some of the issues pointed out on the review process. I have opened issues at https://github.com/daya6489/DriveML with some of my suggestions. The quality of the writing is still a major concern, in my view.

Hi @mirca ,
Please find above the updated version of the paper. We have addressed your comments raised in the couple of GitHub issues.
Please find below my response to each of the comments (my responses are  in italics):

-As part of the on-going JOSS review, could you please update the README and/or related documentations
in order to address how third parties would go about: 1) contributing to the software 2) reporting issues or problems with the software, and 3) seeking support?. -- _Thanks for your comment! We have now added two new files viz., CONTRIBUTING.md and GOVERNANCE.md to address the points you have raised._

- I strongly suggest the removal of the sentence: "These two features would be incorporated in the next release, and we are currently working on it.". The rationale being that the support for these features might not be as straightforward as the authors claim. -- _Thanks for your comment! The concerned sentence is removed from the latest version of the manuscript._

- Instances of "codes" should be written "code" instead.-- _We have corrected it!_

- There are many wrong usages of the symbol "-" throughout the text.- _Thanks for pointing this out! We have replace '-' with ':' wherever appropriate._

- There are still several grammar and punctuation flaws in the text. See this: "it performs various operations such as missing at random features, Date variable transformation, bulk interactions for numerical features, one-hot encoding for categorical variables and finally, feature selection using zero variance, correlation and Area under the curve (AUC) method."--- _Thanks for your comment! We have corrected the issue with the concerned line and have also done a thorough proof-reading to rectify some grammatical inconsistencies._

- I think the citation style is not very adequate specially for entries with many authors. Some lines in the text almost just includes a citation reference. In this case, you should use something like (FIRST_AUTHOR_LAST_NAME et. al. YEAR).-- _Thanks for pointing this out! We have corrected this issue in the latest version of the manuscript._

Overall, we have done another round of proof-reading (both manual and using tools such as, Grammarly) to remove any grammatical inconsistencies.
