### 2020-10-31 #2278 @ledell
@sayanddude I am getting errors running with default settings.

I do not understand why the vignette goes through the effort of demonstrating how to use an "auto-prepped" training set via the `autoDataprep()` function, but then neglects to use the auto-prepped dataset in the examples -- it uses the original `heart` data.frame.  If I try to use the auto-prepped version, I get errors.

Also, why is your vignette called "SmartML.html" when "SmartML" is the name of a different AutoML system that you evaluate in your paper?  I think it would make a lot more sense to call it "DriveML" or some other generic name.

Here's the code to reproduce the error:

```
library(DriveML)

heart <- DriveML::heart

target <- "target_var"
prepdata <- autoDataprep(data = heart, target = target)
heart_prep_master <- prepdata$master_data
model_prep_master <- autoMLmodel(train = heart_prep_master, target = target)
```
Gives this error (I've tried to run it a few times, same error).

```
Error: mtry can not be larger than number of variables in data. Ranger will EXIT now.
Error in ranger::ranger(formula = NULL, dependent.variable.name = tn,  :
  User interrupt or internal error.
In addition: There were 50 or more warnings (use warnings() to see the first 50)
```
