### 2020-10-31 daya6489/DriveML#11 @ledell
[REVIEW-JOSS] Vignette code does not work when using auto-prepped training set
I am getting errors running your vignette code with default settings.

The vignette currently goes the the process of creating an "auto-prepped" training set via the `autoDataprep()` function, but then neglects to use the auto-prepped dataset in the examples -- it uses the original `heart` data.frame.  Is that intentional?  It seems like your package is meant to first use `autoDataprep()` and then `autoMLmodel()` on the auto-prepped data.

If I try to use the auto-prepped version, I get errors.  It seems that there's some problem with how you're setting the range of mtry (it seems to not be actually based on the # of columns of the data, but possibly hard-coded?).

Here's the code to reproduce the error:

```
library(DriveML)

heart <- DriveML::heart

target <- "target_var"
prepdata <- autoDataprep(data = heart, target = target)
heart_prep_master <- prepdata$master_data
model_prep_master <- autoMLmodel(train = heart_prep_master, target = target)
```
Running `autoMLmodel()` on auto-prepped dataset gives this error (I've tried to run it a few times, same error).

```
Error: mtry can not be larger than number of variables in data. Ranger will EXIT now.
Error in ranger::ranger(formula = NULL, dependent.variable.name = tn,  :
  User interrupt or internal error.
In addition: There were 50 or more warnings (use warnings() to see the first 50)
```
And the warnings all look like this:
```
> warnings()
Warning messages:
1: In randomForest.default(m, y, ...) : invalid mtry: reset to within valid range
```

Also, why is your vignette called "SmartML.html" when "SmartML" is the name of a different AutoML system that you evaluate in your paper?  I think it would make a lot more sense to call it "DriveML.html" or some other generic name to avoid any confusion.
