### 2020-10-31 daya6489/DriveML#13 @ledell
[REVIEW-JOSS] Vignette and demo code do not load dataset in the standard way
The standard way to load a dataset that's included in your package is:

```
library(DriveML)
data(heart)
```

In your code demos, you use:
```
heart = DriveML::heart
```

I'd suggest to update the demos and vignettes to use the standard way.
