+++ vignettes/experiment_driveml_large_data.Rmd
@@ -1,0 +1,52 @@
+---
+title: "Experiment of DriveML R package - Large Data set"
+author: "Dayanand, Sayan, Ravi, Kiran R"
+date: "16/06/2021"
+output: html_document
+---
+
+```{r setup0, include=TRUE,eval=FALSE, warning=FALSE}
+library(DriveML)    ## automl package
+library(data.table) ## data wrangling package
+library(caret)      ## ML wrapper package
+library(mlr)        ## ML wrapper package
+library(h2o)        ## automl package
+library(OneR)       ## automl package
+library(Metrics)    ## auc function
+library(autoxgboost)## automl package
+```
+
+```{r setup, include=FALSE, eval=TRUE, warning=FALSE}
+library(DriveML)    ## automl package
+library(data.table) ## data wrangling package
+library(caret)      ## ML wrapper package
+library(mlr)        ## ML wrapper package
+library(h2o)        ## automl package
+library(OneR)       ## automl package
+library(Metrics)    ## auc function
+library(autoxgboost)## automl package
+```
+
+## Adult Data Set
+
+data source : **https://archive.ics.uci.edu/ml/datasets/adult**
+
+#### Content
+
+Extraction was done by Barry Becker from the 1994 Census database. A set of reasonably clean records was extracted using the following conditions: ((AAGE>16) && (AGI>100) && (AFNLWGT>1)&& (HRSWK>0))
+
+Prediction task is to determine whether a person makes over 50K a year.
+
+**Source:**
+
+Donor:
+
+Ronny Kohavi and Barry Becker
+Data Mining and Visualization
+Silicon Graphics.
+e-mail: ronnyk '@' live.com for questions.
+
+##### Problem Statement
+
+Predict whether income exceeds $50K/yr based on census data. Also known as "Census Income" dataset.
+
