+++ vignettes/experiment_driveml_medium_data.Rmd
@@ -1,0 +1,57 @@
+---
+title: "Experiment of DriveML R package - Medium Data set"
+author: "Dayanand, Sayan, Ravi, Kiran R"
+date: "16/06/2021"
+output: html_document
+---
+
+```{r setup0, include=TRUE,eval=FALSE, warning=FALSE}
+library(DriveML)    ## automl package
+library(data.table) ## data wrangling package
+library(caret)      ## ML wrapper package
+library(mlr)        ## ML wrapper package
+library(h2o)        ## automl package
+library(OneR)       ## automl package
+library(Metrics)    ## auc function
+library(autoxgboost)## automl package
+```
+
+```{r setup, include=FALSE, eval=TRUE, warning=FALSE}
+library(DriveML)    ## automl package
+library(data.table) ## data wrangling package
+library(caret)      ## ML wrapper package
+library(mlr)        ## ML wrapper package
+library(h2o)        ## automl package
+library(OneR)       ## automl package
+library(Metrics)    ## auc function
+library(autoxgboost)## automl package
+```
+
+## HR Analytics: Job Change of Data Scientists
+
+data source : **https://www.kaggle.com/arashnic/hr-analytics-job-change-of-data-scientists**
+
+#### Content
+
+A company which is active in Medium data and Data Science wants to hire data scientists among people who successfully pass some courses which conduct by the company. Many people signup for their training. Company wants to know which of these candidates are really wants to work for the company after training or looking for a new employment because it helps to reduce the cost and time as well as the quality of training or planning the courses and categorization of candidates. Information related to demographics, education, experience are in hands from candidates signup and enrollment.
+
+This dataset designed to understand the factors that lead a person to leave current job for HR researches too. By model(s) that uses the current credentials,demographics,experience data you will predict the probability of a candidate to look for a new job or will work for the company, as well as interpreting affected factors on employee decision.
+
+The whole data divided to train and test . Target isnt included in test but the test target values data file is in hands for related tasks. A sample submission correspond to enrollee_id of test set provided too with columns : enrollee _id , target
+
+
+**Note:**
+
+The dataset is imbalanced.
+
+Most features are categorical (Nominal, Ordinal, Binary), some with high cardinality.
+
+Missing imputation can be a part of your pipeline as well.
+
+
+##### Problem Statement
+
+Predict who will move to a new job
+
+Please refer to the following task for more details: **https://www.kaggle.com/arashnic/hr-analytics-job-change-of-data-scientists/tasks?taskId=3015**
+
