### 2020-11-02 daya6489/DriveML#11 @sayanddude
Thank you for pointing this out! We have fixed the issue and corrected the vignette name.
