### 2020-11-02 #2278 @sayanddude
> @sayanddude I am getting errors running your code with default settings. I filed an issue here: [daya6489/DriveML#11](https://github.com/daya6489/DriveML/issues/11)

@ledell Thank you for your comments! I have worked on the issues pointed by you and have closed the Github Issues. I have updated the paper as well with the corrected vignette name/link. Please let me know if I missed anything.
