### 2020-11-02 #2278 @sayanddude
> @sayanddude: When reviewers provide detailed issues, please don't close them with such brusque responses. You should provide a detailed description of what has been fixed and point to the commits where you made those changes (it's definitely not obvious what has happened from the commit history since that seems to get regularly force pushed to remove history). Then you should wait for the **reviewer** to respond that the issue is fixed before closing the issue. It's not up to you to decide if your change was sufficient!

@dfm My apologies! I have edited my post above to mentioned the steps taken to address each of the comments. Also, I have reopened the Github issues.
