### 2020-11-02 #2278 @ledell
@sayanddude Please link to the commits directly in the issues, and have them merged into master before reporting them as "fixed" in the issue.  Right now, I still have to manually look around to find your commits on the "develop" branch to see if the issues were actually solved.
