### 2020-11-03 daya6489/DriveML#13 @daya6489
@ledell Thank you for your comments. Here is the commit link https://github.com/daya6489/DriveML/commit/c2ea6573b4585e2b317fb99f7bdd44584ef3ac14

Please note that we have updated all the changes as suggested by you and some other updates in a single commit yesterday.

We have made the required change in the "README.md" and "DriveML.Rmd" files as shown in the screenshots below and can be seen in the commit link shared above
![image](https://user-images.githubusercontent.com/17849762/97956052-b3625800-1dcd-11eb-84f1-551ebe43becb.png)
![image](https://user-images.githubusercontent.com/17849762/97956162-f8868a00-1dcd-11eb-98bd-6736dd7d7f5a.png)
