### 2020-11-03 daya6489/DriveML#11 @daya6489
@ledell  we have also corrected the vignette name in the paper. Here is the commit link https://github.com/daya6489/DriveML/commit/f4088bc0630d196b0e026432eaf2ce7ca976db23 for the updated paper
