### 2020-11-03 #2278 @sayanddude
> @sayanddude Please link to the commits directly in the issues, and have them merged into master before reporting them as "fixed" in the issue. Right now, I still have to manually look around to find your commits on the "develop" branch to see if the issues were actually solved.

@ledell Sure! We have now mentioned the commit links in each of the open Github Issues. Please let us know if we missed anything. Thanks!
