### 2020-11-26 daya6489/DriveML#15 @ledell
[REVIEW-JOSS] Vignette code does not use auto-prepped dataset in autoMLmodel()
The vignette creates a `train_data` object using the `autoDataprep()` function on the `heart` dataset.  However the main function, `autoMLmodel()` uses the original `heart` dataset instead of the auto-prepped version, `train_data`.

```
mymodel <- autoMLmodel( train = heart,
                        test = NULL,
                        target = 'target_var',
                        testSplit = 0.2,
                        tuneIters = 100,
                        tuneType = "random",
                        models = "all",
                        varImp = 10,
                        liftGroup = 50,
                        maxObs = 4000,
                        uid = NULL,
                        htmlreport = FALSE,
                        seed = 1991)
```
