### 2020-11-26 daya6489/DriveML#16 @ledell
[REVIEW-JOSS] Installation of development version fails on R 3.4
Note that the **prediction** package requires R (≥ 3.5.0), which is a dependency of **iml** (latest version), a package DriveML imports, so if you try to install any of these packages on a machine with R 3.4, it will fail.

Here's some of the errors:

```
ERROR: dependency ‘iml’ is not available for package ‘DriveML’
* removing ‘/home/ledell/R/x86_64-pc-linux-gnu-library/3.4/DriveML’
Error: Command failed (1)
In addition: There were 11 warnings (use warnings() to see them)
> warnings()
Warning messages:
1: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘jsonlite’ had non-zero exit status
2: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘ps’ had non-zero exit status
3: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘rlang’ had non-zero exit status
4: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘processx’ had non-zero exit status
5: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘purrr’ had non-zero exit status
6: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘vctrs’ had non-zero exit status
7: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘callr’ had non-zero exit status
8: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘pillar’ had non-zero exit status
9: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘pkgbuild’ had non-zero exit status
10: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘pkgload’ had non-zero exit status
11: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘iml’ had non-zero exit status
```
