### 2020-11-26 daya6489/DriveML#17 @ledell
[REVIEW-JOSS] Using the `autoDataprep()` function gives worse results than using no data prep
It seems that the `autoDataprep()` function is not doing it's job -- when I run `autoMLmodel()` on two versions of the data (original data vs auto-prepped data), the results are worse for the auto-prepped data.  I don't know if there are some datasets where this is working better, but the vignette example shows degradation in performance after applying `autoDataprep()`:

```
# Compare auto-prep to original data

library(DriveML)

dateprep <- autoDataprep(data = heart,
                         target = 'target_var',
                         missimpute = 'default',
                         auto_mar = FALSE,
                         mar_object = NULL,
                         dummyvar = TRUE,
                         char_var_limit = 15,
                         aucv = 0.002,
                         corr = 0.98,
                         outlier_flag = TRUE,
                         uid = NULL,
                         onlykeep = NULL,
                         drop = NULL)

print(dateprep)
train_data <- dateprep$master_data

mymodel <- autoMLmodel( train = heart,  #original data
                        test = NULL,
                        target = 'target_var',
                        testSplit = 0.2,
                        tuneIters = 100,
                        tuneType = "random",
                        models = "all",
                        varImp = 10,
                        liftGroup = 50,
                        maxObs = 4000,
                        uid = NULL,
                        htmlreport = FALSE,
                        seed = 1991)

mymodel2 <- autoMLmodel( train = train_data,  #autoprepped data
                        test = NULL,
                        target = 'target_var',
                        testSplit = 0.2,
                        tuneIters = 100,
                        tuneType = "random",
                        models = "all",
                        varImp = 10,
                        liftGroup = 50,
                        maxObs = 4000,
                        uid = NULL,
                        htmlreport = FALSE,
                        seed = 1991)

mymodel$results
mymodel2$results
```

Original data, best test AUC is 0.913.

```
> mymodel$results
         Model Fitting time Scoring time Train AUC Test AUC Accuracy Precision Recall F1_score
1       glmnet  22.967 secs   0.011 secs     0.928    0.906    0.820     0.824  0.848    0.836
2       logreg  22.741 secs   0.004 secs     0.929    0.906    0.820     0.824  0.848    0.836
3 randomForest  24.315 secs   0.006 secs     0.979    0.913    0.787     0.778  0.848    0.812
4       ranger  26.385 secs   0.025 secs     1.000    0.902    0.803     0.800  0.848    0.824
5      xgboost  23.323 secs   0.003 secs     0.999    0.890    0.770     0.771  0.818    0.794
6        rpart  22.927 secs   0.005 secs     0.857    0.808    0.787     0.794  0.818    0.806
```
Auto-prepped data, best test AUC is 0.881.

```
> mymodel2$results
         Model Fitting time Scoring time Train AUC Test AUC Accuracy Precision Recall F1_score
1       glmnet  22.458 secs   0.008 secs     0.924    0.864    0.787     0.812  0.788    0.800
2       logreg  22.728 secs   0.004 secs     0.934    0.830    0.770     0.788  0.788    0.788
3 randomForest  25.191 secs   0.016 secs     1.000    0.850    0.787     0.794  0.818    0.806
4       ranger  24.994 secs    0.03 secs     0.992    0.881    0.820     0.824  0.848    0.836
5      xgboost  21.643 secs   0.004 secs     0.994    0.878    0.820     0.806  0.879    0.841
6        rpart  22.469 secs   0.007 secs     0.855    0.805    0.787     0.778  0.848    0.812
```
