### 2020-11-26 #2278 @ledell
# JOSS Review

**Article:** "DriveML: An R Package for Driverless Machine Learning"

**Package URL:** [https://daya6489.github.io/DriveML/](https://daya6489.github.io/DriveML/)

## Summary

There are some key aspects of this article that are missing; it does not meet minimal standards and lacks depth for an article about a new machine learning package.  It currently reads as more of a README or user guide / documentation of DriveML. The package is underdeveloped in important areas -- it currently only implements binary classification and model selection using one metric (test AUC).  It's missing standard tasks such as regression, multiclass classification, does not compute all appropriate performance metrics, even though the package uses the mlr package underneath (which has all of this functionality available).  Lastly, the paper makes the claim that DriveML is "best" when compared to other AutoML packages, which is misleading and unsubstantiated.

High-level review summary:

- DriveML only implements binary clasification, so compared to existing AutoML packages (which offer all supervised learning takss: regression, binary and multiclass classification), it's missing some major functionality.  There doesn't seem to be any practical limitation that would prevent these tasks from being implemented (multiclass is available in mlr, which this is built-on), so I would consider the package to be feature-incomplete in that sense.  There's also no way to execute/run the AutoML algorithm by a time limit (this is a common feature of existing open source AutoML tools).
- There are no benchmarks or experiments on how the AutoML performs (hard to tell if this is a useful ML package in practice, which is an extremely important factor to consider when evaluating whether to use a package).  Without any performance comparisions, all that is presented in the paper is an API and an example of how you'd use a package with that API (which could be ok if that's the stated goal of the paper).  Currently there's no information available to the reader about whether the package produces high quality results, competitive with AutoML packages in R, Python, etc.  In order to be a useful AutoML package in R, it's important that the models produced by the package are better (or at least equal) than what someone would get from a simple grid/random search in mlr or caret.
- The auto-data prep function, `autoDataprep()`, actually hurts performance (on the data provided in the vignette) when used (as directed in the vignette).  There's a lack of testing around performance and utility of algorithms in the package.
- Comparison to existing AutoML software in R is limited to a checklist of which existing AutoML packages share some of DriveML's features.  Claims that DriveML performs "best" is based on this cherry-picked list of features, which does not provide any valid grounds for this claim.  A feature table is useful, but it should be complete and not favor the functionality of a single package.  It is not valid to claim a AutoML package is "best" based on a simple list of features, rather than the performance of the actual AutoML algorithms that are provided by such packages.
- Vingette is not yet at a "publication quality" level.  There are a number of functions included in the package that do not appear in the vignette.  There's a number of grammar and communication issues (e.g. _"Frequent tranformer for categorical features"_, _"Mutliple binary classification included like logistic regression, randomForest, xgboost, glmnet, rpart"_).  There are also basic formatting issues with code indentation and spacing.

## Detailed Review

### Title

Name of the paper suggested change from:

*"DriveML: An R Package for Driverless Machine Learning"*

to:

*"DriveML: An R Package for Automated Machine Learning"* or *DriveML: Self-Drive Machine Learning Projects"*

The vignette and package name both use a different subtitle, *"Self-Drive Machine Learning Projects"*.  Here are a few of the drawbacks of using the term "Driverless" in the subtitle:

- This is not a techincal term, and "Automated Machine Learning" is a well-understood, accurate description of what type of software it is.
- It's both "Drive" (DriveML) and "Driverless", which is a bit contradictory.
- The term "Driverless" is the name of a popular, different AutoML software tool (["Driverless AI"](https://www.h2o.ai/products/h2o-driverless-ai/) by H2O.ai).

Since the package only seems to support binary classification problems, it might be more appropriate to note that in the title; a more apropriate subtitle would be something like:  *"DriveML: An R Package for Automated Binary Classification"*

### Abstract (Summary)

> DriveML performs the best across different features.

This is vague/unclear claim (at best) and incorrect claim (at worst).  The word "performs" suggests that DriveML was compared to a set of AutoML tools based on the performance of the package (i.e. model performance).  However, the paper compares a list of DriveML features and indicates whether other packages implement that feature (DriveML declared best, based on this).  See Table 1 section below for more feedback.

> Overall, the main benefits of DriveML are in development time savings, reduce developer’s errors, optimal tuning of machine learning models, and reproducibility

All AutoML tools aim to reduce time spent by the developer and minimize errors, this is not a contribution unique to DriveML (should be reworded to indicate that this is true of all AutoML tools, or removed).

There's no evidence that the models are "optimal" (optimal compared to what?), so this is not an appropriate term to use.

The abstract should highlight any unique features that DriveML offers over the existing AutoML packages in R.

There's no clear mentioning of what types of tasks this package performs.  Most machine learning packages (mlr, caret, randomForest, xgboost, h2o, etc) can do regression, binary classification and multiclass classification.  The expectation is that any ML package should do all three of these, especially a package that builds on a package like mlr, which already supports all the tasks.  It should be made very clear that the only task supported with this package is binary classification.

### Key Functionality

Key functionality that is not explained: how to run this (how long does it run?).  Most AutoML software has the ability to limit the run by time, and without such a feature, it makes it very difficult to compare the performance of one AutoML system to another.  The `autoMLmodel()` function has the `tuneIters` argument, but there's no explaination of how that works in the paper or vignette, or how an "iteration
" relates to the runtime the AutoML algorithm.  There's no guidance on how to specify this argument.  The only thing a user can assume is that more iterations will lead to a longer runtime.

> b) Performs model validation usingReceiver Operating Characteristics Area Under the Curve (ROC AUC)

Is there a reason that the package only optimzes for AUC in the model validation process?

> b) Performs model validation using Receiver Operating Characteristics Area Under the Curve (ROC AUC),

Can the user change the metric used?  If not, this should be acknowledged as a limitation.

> The other features of the autoMLmodel function are (a) Hyperparameter Tuning: Random search (default option)

There's very little information provided about the tuning.  How are the parameters chosen to search over and what values are searched over?  Is that hardcoded somewhere?  Table 1 says "Tuning" is performed, but it's also not clear what data is used to tune the models.

It appears as though the test set is used to "tune" each algorithm type (e.g. Random Forest), and then the best model from each type is returned onto a grid of results?

Here's the Random Forest data:

```
> small_ml_random$trainedModels$randomForest$tuneData
HyperParsEffectData:
Hyperparameters: mtry,ntree,importance,nodesize
Measures: auc.test.mean
Optimizer: TuneControlRandom
Nested CV Used: FALSE
[1] "Partial dependence requested"
Snapshot of data:
  mtry ntree importance nodesize auc.test.mean iteration exec.time
1    4   140      FALSE       14     0.8905879         1     0.036
2    7   174      FALSE       18     0.8726435         2     0.025
3    6   476       TRUE       15     0.8709677         3     0.051
4    5    84       TRUE        2     0.8810920         4     0.021
5    7   533      FALSE       12     0.8782991         5     0.061
  threshold
1 0.9999339
2 0.9999339
3 0.9999339
4 0.9999339
5 0.9999339
```

The best `auc.test.mean` value is 0.89, but then it's confusing because that number doesn't match up what's in the `results` section:

```
> small_ml_random$results
         Model Fitting time Scoring time Train AUC Test AUC
1       glmnet   1.387 secs   0.007 secs     0.928    0.865
3 randomForest   1.153 secs   0.008 secs     0.983    0.864
4       ranger   1.178 secs   0.014 secs     0.999    0.845
5      xgboost   1.037 secs   0.004 secs     0.970    0.824
2       logreg   1.011 secs   0.005 secs     0.963    0.765
6        rpart   1.068 secs   0.006 secs     0.963    0.734
  Accuracy Precision Recall F1_score
1    0.770     0.788  0.788    0.788
3    0.787     0.794  0.818    0.806
4    0.770     0.788  0.788    0.788
5    0.721     0.735  0.758    0.746
2    0.738     0.758  0.758    0.758
6    0.672     0.697  0.697    0.697
```

> Apart from this, DriveML can call the SmartEDA (Putatunda et al., 2019) package for performing automated exploratory data analysis (EDA).

It's not appropriate to say that "DriveML can call the SmartEDA package", since they are independent packages (one does not call the other).  Users can use the SmartEDA package or any other EDA package to do EDA.

There's no `print()` functionality for the "autoMLmodel" objects.  If you return or print the object, it will spill a lot of text with all the information, which is not very user friendly.

#### Table 1

Many problems with this table:

OneR is not an AutoML package.  It's a package that implements a simple algorithm called the "One Rule Machine Learning Classification Algorithm".  This is no more an AutoML package than any other single algorithm package in R (e.g. rpart, randomForest, xgboost, etc).  It should not be included in this list.

List of AutoML tools is limited.  Other R packages for AutoML include: [RemixAutoML](https://github.com/AdrianAntico/RemixAutoML), [autoxgboost](https://github.com/ja-thomas/autoxgboost) and [autokeras](https://github.com/r-tensorflow/autokeras).  They should be mentioned for completeness.

The names of the packages are not correct (they should use proper capitalization since R package names are case-sensitive. e.g. "h2o", "automl", ).

**Major issue:**  The list of "desired characteristics" is effectively a list of DriveML's features and then a checkbox for whether another package also has that feature.  However, in the paper, it's presented as a comprehensive list of features that an AutoML package could have.  From looking at this list, it appears that DriveML has the most unique features out of any of the AutoML packages, which is not true by any means.  And from this, there's a big leap taken to get to an statement like this (from the abstract):

> DriveML performs the best across different features.

In the table, the "Fast EDA" task is counted as part of DriveML's features, but all the EDA is from SmartEDA, a separate package.  The SmartEDA was published in a JOSS article, so that functionality should not be counted as part of DriveML's offerings.

> *"None of the other packages provide these functionalities"*.

That's not true, at least in the case of H2O, which has (f) Tuning, (i) PD Plots (there's a typo of PDP Plots in the paper), (k) HTML report using rmarkdown, in addition to many other features that are not listed in Table 1. There's a huge list of features that H2O AutoML has that DriveML does not have (for example), but that is not shown here.  It would be easy to add 15 features to Table 1 that H2O has and DriveML does not.

It's not useful to the reader to see a biased view of "desired characteristics" that closely match the DriveML feature list.  It's misleading and frames DriveML as the most feature complete AutoML system in R, which is not accurate.

Most importantly, it's not appropriate to claim that "DriveML performs the best across different features".  This is a vague claim ("best performance"?) and the list of features it's comparing against is a biased subset of all AutoML potential features that happens to line up very closely with the current list of DriveML features.

There is absolutely no basis to claim that DriveML is "best" by any measure, based on what is presented in this paper.  A realistic way to compare AutoML packages is to perform a comprehensive study and benchmark across a large number of carefully chosen datasets, such as the [OpenML AutoML Benchmark](https://github.com/openml/automlbenchmark) (2019 ICML AutoML Workshop).

> Thus, Table 1 shows that the DriveML package is better than most of the other competing packages such as OneR, H2O, SmartML and AutoML.

This is not true at all.  This should be removed, along with the same claim in the abstract.

#### Functionality not discussed

The model output (and plots) are buried inside the "autoMLmodel" object and there's no mention of how to use them in the paper.

```
Threshold <- mymodel$trainedModels$randomForest$modelPlots$Threshold
```

There are no automatic or easy functions to see these plots and compare several models at once (which is a typical functionality of AutoML packages).

### Package Readme.md

All of the EDA tools are from a different package (by same author), "SmartEDA", and it's not clear why that's being presented as part of the DriveML package.   The documentation does not clarify that these functions are not from the DriveML package, but a different package, so if they are to be kept in the README, that should be clarified.

There were errors running some of the SmartEDA code:

```
> ## Generate Boxplot by category
> ExpNumViz(heart, gp = "target_var", type = 2, nlim = 25, Page = c(2, 2))
Error in ExpNumViz(heart, gp = "target_var", type = 2, nlim = 25, Page = c(2,  :
  unused argument (gp = "target_var")
>
> ## Generate Density plot
> ExpNumViz(heart, gp = NULL, type = 3, nlim = 10, Page = c(2, 2))
Error in ExpNumViz(heart, gp = NULL, type = 3, nlim = 10, Page = c(2,  :
  unused argument (gp = NULL)
```

More errors:

```
> autoMLReport(mlobject = small_ml_random, mldata = small_data, op_file = "driveML_ouput_heart_data.html")
Error in parse_block(g[-1], g[1], params.src, markdown_mode) :
  Duplicate chunk label 'mdp131', which has been used for the chunk:
cc  <- mldata$raw_list$AUC_value

if(nrow(cc) > 0){
  rownames(cc) <- NULL
  cc <- cc[, c(2:3)]
}

paged_table(cc)
```

### Code & Vignettes

```
devtools::install_github("daya6489/DriveML")
```

The command to install from github fails on a clean machine -- there should be some extra commands to install the dependencies so there's not so much debugging to get the package to install.

Here's some of the errors:

```
ERROR: dependency ‘iml’ is not available for package ‘DriveML’
* removing ‘/home/ledell/R/x86_64-pc-linux-gnu-library/3.4/DriveML’
Error: Command failed (1)
In addition: There were 11 warnings (use warnings() to see them)
> warnings()
Warning messages:
1: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘jsonlite’ had non-zero exit status
2: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘ps’ had non-zero exit status
3: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘rlang’ had non-zero exit status
4: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘processx’ had non-zero exit status
5: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘purrr’ had non-zero exit status
6: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘vctrs’ had non-zero exit status
7: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘callr’ had non-zero exit status
8: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘pillar’ had non-zero exit status
9: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘pkgbuild’ had non-zero exit status
10: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘pkgload’ had non-zero exit status
11: In utils::install.packages(pkgs, repos = repos, type = type,  ... :
  installation of package ‘iml’ had non-zero exit status
```

~~Why is your vignette called "SmartML.html" when "SmartML" is the name of a different AutoML system that you evaluate in your paper? I think it would make a lot more sense to call it "DriveML.html" or some other generic name to avoid any confusion.~~
Update: This was addressed [here](https://github.com/daya6489/DriveML/commit/f39fa73c90cd29cea3168719f55ffda732086d9f).

##### Section 3.1: Data Exploration:

Vignette has a whole section on "Data Exploration" (3.1) which is actually just a tour of the functionality of the SmartEDA package.  The SmartEDA demo could be referenced, but it should not be included as part of the DriveML vignette. Currently, the Data Exploration section demos seven SmartEDA functions, and so it's a long vignette already, even before you get to the DriveML functionality.  Vignettes should be include a minimal example of all the functionality of the package, and it's better to keep the SmartEDA demo inside it's own vignette.

##### Section 3.2: Data Preparations using `autoDataprep`:

- The code in this section is not formatted correctly in the HTML knitted doc (indentation levels are inconsistent in the demo code).
- The `autoDataprep()` function is supposed to be automatic, but the demo shows setting a large number of arguments with specific values.  If those are the default values, then they do not need to be shown becuase it's not clear whether they are required values or all have defaults.
- In the `print(dateprep)` output (the `complete_data` section), there's 303 rows printed out in the vigenette.  The function should not print that much data -- it's advised to only print 10-20 rows in a vignette (or in any demo / document).  That should be fixed.

I do not understand why the vignette goes through the effort of demonstrating how to create an "auto-prepped" training set via the `autoDataprep()` function, but then neglects to use the auto-prepped dataset in the examples -- it uses the original `heart` data.frame.

There does not seem to be any demonstration of how to apply the data prep to the test set.  This should be shown in the vignette.  It seems that it might be the `predictDataprep()` function that needs to be added to the vignette.

If you compare the AutoML performance to when you use the auto-prepped dataset to when you don't, it actually loses performance when you do the auto-prepping.  This seems to actually hurt performance on the dataset given in the vignette, as well as on other datasets that I tried.

##### Section 3.3: Machine learning models usuing `autoMLmodel`:

- When you say "Logistic regression - logreg" and then have five algorithms indendented as a "subset" of that, it sounds like you're saying there are five types of "logistic regression".
- The code in this section is not formatted correctly in the HTML knitted doc (extra space before the `train` arg in the `autoMLmodel()` function demo.
[…]
