### 2020-11-30 #2278 @dfm
@ledell: Thank you very much for your thorough and detailed review!

Given the significant major issues that are still present in this submission I'm going to loop in @openjournals/joss-eics (especially @arfon) for their perspective.
