### 2020-12-07 #2278 @sayanddude
> @ledell - many thanks for your detailed review here. Your efforts are most appreciated (as are @pat-s' and @mirca's).
>
> @sayanddude - It's clear that this package still requires substantial work to pass the review here, and given that this submission has been flagged as requiring substantial work (`pending-major-enhancements`) for some time, at this point I believe it is most appropriate to reject the submission.
>
> My primary concern here is that this submission is not [meeting our criteria around contribution to the field](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). In particular we say:
>
> > In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools). "Minor utility" packages, including "thin" API clients, and single-function packages are not acceptable.
>
> @ledell's review has convinced me (and @dfm) that this package still requires significant work.
>
> Thank you to all of the reviewers, and to you @sayanddude for working with the feedback given. If you are able to substantially improve this submission, a re-submission to JOSS in the future would be welcome.

Thanks @arfon!  I understand the decision taken and would definitely consider JOSS for submission in the future as I believe that this journal is a step in the right direction when compared to the review process in some of the top scopus indexed journals or the ones with high impact factor!
Just one feedback- It would be great if the reviewers are taken only if they can comment on the work done by the authors after their feedback. For example in this case, after the review was done by the 2nd reviewer, he could not continue but still we did worked on the comments given. So if a reviewer is going to be changed mid-way, then either the same comments need to be followed up by another reviewer or those comments should be made null and void to avoid unnecessary wastage of time and effort.
